# apps/api/internal/httpapi/correspondence/generated.go
// Package correspondence provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package correspondence

import (
	"context"
	"fmt"
	"time"

	"github.com/gofiber/fiber/v3"
	"github.com/oapi-codegen/runtime"
)

// Defines values for DeliveryStatus.
const (
	Failed   DeliveryStatus = "failed"
	Pending  DeliveryStatus = "pending"
	Prepared DeliveryStatus = "prepared"
	Sent     DeliveryStatus = "sent"
	Unknown  DeliveryStatus = "unknown"
)

// Valid indicates whether the value is a known member of the DeliveryStatus enum.
func (e DeliveryStatus) Valid() bool {
	switch e {
	case Failed:
		return true
	case Pending:
		return true
	case Prepared:
		return true
	case Sent:
		return true
	case Unknown:
		return true
	default:
		return false
	}
}

// Defines values for SendInputSendMode.
const (
	Separate SendInputSendMode = "separate"
	Shared   SendInputSendMode = "shared"
)

// Valid indicates whether the value is a known member of the SendInputSendMode enum.
func (e SendInputSendMode) Valid() bool {
	switch e {
	case Separate:
		return true
	case Shared:
		return true
	default:
		return false
	}
}

// Delivery defines model for Delivery.
type Delivery struct {
	Error              *string        `json:"error,omitempty"`
	GmailMessageId     *string        `json:"gmailMessageId,omitempty"`
	GmailThreadId      *string        `json:"gmailThreadId,omitempty"`
	Id                 string         `json:"id"`
	Recipients         []string       `json:"recipients"`
	ReplyAllRecipients []string       `json:"replyAllRecipients"`
	RfcMessageId       *string        `json:"rfcMessageId,omitempty"`
	Status             DeliveryStatus `json:"status"`
}

// DeliveryStatus defines model for Delivery.Status.
type DeliveryStatus string

// Draft defines model for Draft.
type Draft struct {
	Attachments      *[]string `json:"attachments,omitempty"`
	Bcc              []string  `json:"bcc"`
	Cc               []string  `json:"cc"`
	Html             string    `json:"html"`
	Id               string    `json:"id"`
	MailboxId        string    `json:"mailboxId"`
	ReplyToMessageId *string   `json:"replyToMessageId,omitempty"`
	Subject          string    `json:"subject"`
	ThreadId         *string   `json:"threadId,omitempty"`
	To               []string  `json:"to"`
	TrackConversions *bool     `json:"trackConversions,omitempty"`
	TrackLinks       *bool     `json:"trackLinks,omitempty"`
	TrackedDocuments *[]string `json:"trackedDocuments,omitempty"`
	UpdatedAt        time.Time `json:"updatedAt"`
}

// DraftInput defines model for DraftInput.
type DraftInput struct {
	Attachments      *[]string `json:"attachments,omitempty"`
	Bcc              []string  `json:"bcc"`
	Cc               []string  `json:"cc"`
	Html             string    `json:"html"`
	Id               *string   `json:"id,omitempty"`
	MailboxId        string    `json:"mailboxId"`
	ReplyToMessageId *string   `json:"replyToMessageId,omitempty"`
	Subject          string    `json:"subject"`
	ThreadId         *string   `json:"threadId,omitempty"`
	To               []string  `json:"to"`
	TrackConversions *bool     `json:"trackConversions,omitempty"`
	TrackLinks       *bool     `json:"trackLinks,omitempty"`
	TrackedDocuments *[]string `json:"trackedDocuments,omitempty"`
}

// SendInput defines model for SendInput.
type SendInput struct {
	Draft          DraftInput        `json:"draft"`
	IdempotencyKey string            `json:"idempotencyKey"`
	SendMode       SendInputSendMode `json:"sendMode"`
}

// SendInputSendMode defines model for SendInput.SendMode.
type SendInputSendMode string

// SendResult defines model for SendResult.
type SendResult struct {
	ConversationId string     `json:"conversationId"`
	Deliveries     []Delivery `json:"deliveries"`
}

// SaveDraftJSONRequestBody defines body for SaveDraft for application/json ContentType.
type SaveDraftJSONRequestBody = DraftInput

// SendTrackedMessageJSONRequestBody defines body for SendTrackedMessage for application/json ContentType.
type SendTrackedMessageJSONRequestBody = SendInput

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (GET /drafts)
	ListDrafts(c fiber.Ctx) error

	// (POST /drafts)
	SaveDraft(c fiber.Ctx) error

	// (DELETE /drafts/{draftId})
	DeleteDraft(c fiber.Ctx, draftId string) error

	// (GET /drafts/{draftId})
	GetDraft(c fiber.Ctx, draftId string) error

	// (POST /send)
	SendTrackedMessage(c fiber.Ctx) error
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []HandlerMiddlewareFunc
}

type MiddlewareFunc fiber.Handler
type HandlerMiddlewareFunc func(c fiber.Ctx, next fiber.Handler) error

// ListDrafts operation middleware
func (siw *ServerInterfaceWrapper) ListDrafts(c fiber.Ctx) error {

	handler := func(c fiber.Ctx) error {
		return siw.Handler.ListDrafts(c)
	}

	for i := len(siw.HandlerMiddlewares) - 1; i >= 0; i-- {
		m := siw.HandlerMiddlewares[i]
		next := handler
		handler = func(c fiber.Ctx) error {
			return m(c, next)
		}
	}

	return handler(c)
}

// SaveDraft operation middleware
func (siw *ServerInterfaceWrapper) SaveDraft(c fiber.Ctx) error {

	handler := func(c fiber.Ctx) error {
		return siw.Handler.SaveDraft(c)
	}

	for i := len(siw.HandlerMiddlewares) - 1; i >= 0; i-- {
		m := siw.HandlerMiddlewares[i]
		next := handler
		handler = func(c fiber.Ctx) error {
			return m(c, next)
		}
	}

	return handler(c)
}

// DeleteDraft operation middleware
func (siw *ServerInterfaceWrapper) DeleteDraft(c fiber.Ctx) error {

	var err error
	_ = err

	// ------------- Path parameter "draftId" -------------
	var draftId string

	err = runtime.BindStyledParameterWithOptions("simple", "draftId", c.Params("draftId"), &draftId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		return fiber.NewError(fiber.StatusBadRequest, fmt.Errorf("Invalid format for parameter draftId: %w", err).Error())
	}

	handler := func(c fiber.Ctx) error {
		return siw.Handler.DeleteDraft(c, draftId)
	}

	for i := len(siw.HandlerMiddlewares) - 1; i >= 0; i-- {
		m := siw.HandlerMiddlewares[i]
		next := handler
		handler = func(c fiber.Ctx) error {
			return m(c, next)
		}
	}

	return handler(c)
}

// GetDraft operation middleware
func (siw *ServerInterfaceWrapper) GetDraft(c fiber.Ctx) error {

	var err error
	_ = err

	// ------------- Path parameter "draftId" -------------
	var draftId string

	err = runtime.BindStyledParameterWithOptions("simple", "draftId", c.Params("draftId"), &draftId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		return fiber.NewError(fiber.StatusBadRequest, fmt.Errorf("Invalid format for parameter draftId: %w", err).Error())
	}

	handler := func(c fiber.Ctx) error {
		return siw.Handler.GetDraft(c, draftId)
	}

	for i := len(siw.HandlerMiddlewares) - 1; i >= 0; i-- {
		m := siw.HandlerMiddlewares[i]
		next := handler
		handler = func(c fiber.Ctx) error {
			return m(c, next)
		}
	}

	return handler(c)
}

// SendTrackedMessage operation middleware
func (siw *ServerInterfaceWrapper) SendTrackedMessage(c fiber.Ctx) error {

	handler := func(c fiber.Ctx) error {
		return siw.Handler.SendTrackedMessage(c)
	}

	for i := len(siw.HandlerMiddlewares) - 1; i >= 0; i-- {
		m := siw.HandlerMiddlewares[i]
		next := handler
		handler = func(c fiber.Ctx) error {
			return m(c, next)
		}
	}

	return handler(c)
}

// FiberServerOptions provides options for the Fiber server.
type FiberServerOptions struct {
	BaseURL            string
	Middlewares        []MiddlewareFunc
	HandlerMiddlewares []HandlerMiddlewareFunc
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router fiber.Router, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, FiberServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router fiber.Router, si ServerInterface, options FiberServerOptions) {
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.HandlerMiddlewares,
	}

	for _, m := range options.Middlewares {
		router.Use(fiber.Handler(m))
	}

	router.Get(options.BaseURL+"/drafts", wrapper.ListDrafts)

	router.Post(options.BaseURL+"/drafts", wrapper.SaveDraft)

	router.Delete(options.BaseURL+"/drafts/:draftId", wrapper.DeleteDraft)

	router.Get(options.BaseURL+"/drafts/:draftId", wrapper.GetDraft)

	router.Post(options.BaseURL+"/send", wrapper.SendTrackedMessage)

}

type ListDraftsRequestObject struct {
}

type ListDraftsResponseObject interface {
	VisitListDraftsResponse(ctx fiber.Ctx) error
}

type ListDrafts200JSONResponse []Draft

func (response ListDrafts200JSONResponse) VisitListDraftsResponse(ctx fiber.Ctx) error {
	ctx.Response().Header.Set("Content-Type", "application/json")
	ctx.Status(200)

	return ctx.JSON(&response)
}

type SaveDraftRequestObject struct {
	Body *SaveDraftJSONRequestBody
}

type SaveDraftResponseObject interface {
	VisitSaveDraftResponse(ctx fiber.Ctx) error
}

type SaveDraft200JSONResponse Draft

func (response SaveDraft200JSONResponse) VisitSaveDraftResponse(ctx fiber.Ctx) error {
	ctx.Response().Header.Set("Content-Type", "application/json")
	ctx.Status(200)

	return ctx.JSON(&response)
}

type DeleteDraftRequestObject struct {
	DraftId string `json:"draftId"`
}

type DeleteDraftResponseObject interface {
	VisitDeleteDraftResponse(ctx fiber.Ctx) error
}

type DeleteDraft204Response struct {
}

func (response DeleteDraft204Response) VisitDeleteDraftResponse(ctx fiber.Ctx) error {
	ctx.Status(204)
	return nil
}

type GetDraftRequestObject struct {
	DraftId string `json:"draftId"`
}

type GetDraftResponseObject interface {
	VisitGetDraftResponse(ctx fiber.Ctx) error
}

type GetDraft200JSONResponse Draft

func (response GetDraft200JSONResponse) VisitGetDraftResponse(ctx fiber.Ctx) error {
	ctx.Response().Header.Set("Content-Type", "application/json")
	ctx.Status(200)

	return ctx.JSON(&response)
}

type SendTrackedMessageRequestObject struct {
	Body *SendTrackedMessageJSONRequestBody
}

type SendTrackedMessageResponseObject interface {
	VisitSendTrackedMessageResponse(ctx fiber.Ctx) error
}

type SendTrackedMessage200JSONResponse SendResult

func (response SendTrackedMessage200JSONResponse) VisitSendTrackedMessageResponse(ctx fiber.Ctx) error {
	ctx.Response().Header.Set("Content-Type", "application/json")
	ctx.Status(200)

	return ctx.JSON(&response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {

	// (GET /drafts)
	ListDrafts(ctx context.Context, request ListDraftsRequestObject) (ListDraftsResponseObject, error)

	// (POST /drafts)
	SaveDraft(ctx context.Context, request SaveDraftRequestObject) (SaveDraftResponseObject, error)

	// (DELETE /drafts/{draftId})
	DeleteDraft(ctx context.Context, request DeleteDraftRequestObject) (DeleteDraftResponseObject, error)

	// (GET /drafts/{draftId})
	GetDraft(ctx context.Context, request GetDraftRequestObject) (GetDraftResponseObject, error)

	// (POST /send)
	SendTrackedMessage(ctx context.Context, request SendTrackedMessageRequestObject) (SendTrackedMessageResponseObject, error)
}

type StrictHandlerFunc func(ctx fiber.Ctx, args any) (any, error)

type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// ListDrafts operation middleware
func (sh *strictHandler) ListDrafts(ctx fiber.Ctx) error {
	var request ListDraftsRequestObject

	handler := func(ctx fiber.Ctx, request interface{}) (interface{}, error) {
		return sh.ssi.ListDrafts(ctx.Context(), request.(ListDraftsRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "ListDrafts")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(ListDraftsResponseObject); ok {
		if err := validResponse.VisitListDraftsResponse(ctx); err != nil {
			return err
		}
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// SaveDraft operation middleware
func (sh *strictHandler) SaveDraft(ctx fiber.Ctx) error {
	var request SaveDraftRequestObject

	var body SaveDraftJSONRequestBody
	if err := ctx.Bind().Body(&body); err != nil {
		return fiber.NewError(fiber.StatusBadRequest, err.Error())
	}
	request.Body = &body

	handler := func(ctx fiber.Ctx, request interface{}) (interface{}, error) {
		return sh.ssi.SaveDraft(ctx.Context(), request.(SaveDraftRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "SaveDraft")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(SaveDraftResponseObject); ok {
		if err := validResponse.VisitSaveDraftResponse(ctx); err != nil {
			return err
		}
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// DeleteDraft operation middleware
func (sh *strictHandler) DeleteDraft(ctx fiber.Ctx, draftId string) error {
	var request DeleteDraftRequestObject

	request.DraftId = draftId

	handler := func(ctx fiber.Ctx, request interface{}) (interface{}, error) {
		return sh.ssi.DeleteDraft(ctx.Context(), request.(DeleteDraftRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "DeleteDraft")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(DeleteDraftResponseObject); ok {
		if err := validResponse.VisitDeleteDraftResponse(ctx); err != nil {
			return err
		}
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// GetDraft operation middleware
func (sh *strictHandler) GetDraft(ctx fiber.Ctx, draftId string) error {
	var request GetDraftRequestObject

	request.DraftId = draftId

	handler := func(ctx fiber.Ctx, request interface{}) (interface{}, error) {
		return sh.ssi.GetDraft(ctx.Context(), request.(GetDraftRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetDraft")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetDraftResponseObject); ok {
		if err := validResponse.VisitGetDraftResponse(ctx); err != nil {
			return err
		}
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// SendTrackedMessage operation middleware
func (sh *strictHandler) SendTrackedMessage(ctx fiber.Ctx) error {
	var request SendTrackedMessageRequestObject

	var body SendTrackedMessageJSONRequestBody
	if err := ctx.Bind().Body(&body); err != nil {
		return fiber.NewError(fiber.StatusBadRequest, err.Error())
	}
	request.Body = &body

	handler := func(ctx fiber.Ctx, request interface{}) (interface{}, error) {
		return sh.ssi.SendTrackedMessage(ctx.Context(), request.(SendTrackedMessageRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "SendTrackedMessage")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(SendTrackedMessageResponseObject); ok {
		if err := validResponse.VisitSendTrackedMessageResponse(ctx); err != nil {
			return err
		}
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}