# packages/api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: Mail Tracker API
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - owner: []
paths:
...
  /drafts:
    get:
      tags: [correspondence]
      operationId: listDrafts
      responses:
        '200':
          description: saved drafts
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Draft' } }
    post:
      tags: [correspondence]
      operationId: saveDraft
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DraftInput' }
      responses:
        '200':
          description: saved draft
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Draft' }
  /drafts/{draftId}:
    get:
      tags: [correspondence]
      operationId: getDraft
      parameters:
        - { name: draftId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: draft
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Draft' }
    delete:
      tags: [correspondence]
      operationId: deleteDraft
      parameters:
        - { name: draftId, in: path, required: true, schema: { type: string } }
      responses:
        '204': { description: deleted }
  /send:
    post:
      tags: [correspondence]
      operationId: sendTrackedMessage
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SendInput' }
      responses:
        '200':
          description: Persisted send attempt. New deliveries start pending and update in the background; repeat idempotency keys return current outcomes.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SendResult' }
...
components:
  securitySchemes:
    owner: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
...
    DraftInput:
      type: object
      required: [mailboxId, to, cc, bcc, subject, html]
      properties:
        trackConversions: { type: boolean, default: false }
        trackLinks: { type: boolean, default: false }
        trackedDocuments: { type: array, maxItems: 10, items: { type: string } }
        id: { type: string }
        mailboxId: { type: string }
        to: { type: array, items: { type: string } }
        cc: { type: array, items: { type: string } }
        bcc: { type: array, items: { type: string } }
        subject: { type: string }
        html: { type: string }
        replyToMessageId: { type: string }
        threadId: { type: string }
        attachments: { type: array, items: { type: string } }
    Draft:
      allOf:
        - { $ref: '#/components/schemas/DraftInput' }
        - type: object
          required: [id, updatedAt]
          properties:
            id: { type: string }
            updatedAt: { type: string, format: date-time }
    SendInput:
      type: object
      required: [draft, sendMode, idempotencyKey]
      properties:
        draft: { $ref: '#/components/schemas/DraftInput' }
        sendMode: { type: string, enum: [shared, separate] }
        idempotencyKey: { type: string }
    SendResult:
      type: object
      required: [conversationId, deliveries]
      properties:
        conversationId: { type: string }
        deliveries: { type: array, items: { $ref: '#/components/schemas/Delivery' } }
    Delivery:
      type: object
      required: [id, recipients, replyAllRecipients, status]
      properties:
        id: { type: string }
        recipients: { type: array, items: { type: string } }
        replyAllRecipients: { type: array, items: { type: string } }
        status: { type: string, enum: [pending, sent, failed, unknown, prepared] }
        error: { type: string }
        gmailMessageId: { type: string }
        gmailThreadId: { type: string }
        rfcMessageId: { type: string }
...